fix: keep nav-and-bg preview page server-rendered
